The mutexes work when declared inside the rw_lock_t struct.
This is an UGLY hack to initialize the mutex and the condition variables.
Malloc can not work since the mutex/cond variables need to be initialized
at creation time and not later.
The solution is to statically initialize the whole struct.
WARNING: This is heavily dependant on the struct declarations. If the
declaration order changes, IT WILL NOT WORK!